Create options class
